introduce CommandMenuContainer
