+++ packages/twenty-front/src/modules/command-menu/hooks/useCommandMenuHotKeys.ts
@@ -2,7 +2,6 @@ import { useCommandMenu } from '@/command-menu/hooks/useCommandMenu';
 import { commandMenuSearchState } from '@/command-menu/states/commandMenuSearchState';
 import { contextStoreNumberOfSelectedRecordsComponentState } from '@/context-store/states/contextStoreNumberOfSelectedRecordsComponentState';
 import { contextStoreTargetedRecordsRuleComponentState } from '@/context-store/states/contextStoreTargetedRecordsRuleComponentState';
-import { useKeyboardShortcutMenu } from '@/keyboard-shortcut-menu/hooks/useKeyboardShortcutMenu';
 import { useScopedHotkeys } from '@/ui/utilities/hotkey/hooks/useScopedHotkeys';
 import { AppHotkeyScope } from '@/ui/utilities/hotkey/types/AppHotkeyScope';
 import { useSetRecoilComponentStateV2 } from '@/ui/utilities/state/component-state/hooks/useSetRecoilComponentStateV2';
@@ -11,10 +10,9 @@ import { useRecoilValue } from 'recoil';
 import { Key } from 'ts-key-enum';
 
 export const useCommandMenuHotKeys = () => {
-  const { toggleCommandMenu, closeCommandMenu } = useCommandMenu();
+  const { closeCommandMenu } = useCommandMenu();
 
   const commandMenuSearch = useRecoilValue(commandMenuSearchState);
-  const { closeKeyboardShortcutMenu } = useKeyboardShortcutMenu();
 
   const setContextStoreTargetedRecordsRule = useSetRecoilComponentStateV2(
     contextStoreTargetedRecordsRuleComponentState,
@@ -24,16 +22,6 @@ export const useCommandMenuHotKeys = () => {
   const setContextStoreNumberOfSelectedRecords = useSetRecoilComponentStateV2(
     contextStoreNumberOfSelectedRecordsComponentState,
     'command-menu',
-  );
-
-  useScopedHotkeys(
-    'ctrl+k,meta+k',
-    () => {
-      closeKeyboardShortcutMenu();
-      toggleCommandMenu();
-    },
-    AppHotkeyScope.CommandMenu,
-    [toggleCommandMenu],
   );
 
   useScopedHotkeys(
